requirements.txt
functions_framework
google-cloud-storage
google-cloud-firestore
